Update and rename blog-post.js to blogPost.js
